Move `logs` command to `services logs` for consistency
The logs command is now namespaced under services to be consistent with
other service-related commands (start, stop, restart, status).

Usage change: `denvig logs <name>` → `denvig services logs <name>`
